move new traits into separate file
